fix: form.test file's extension and tests
